Repaso de Repositorios y Servicios

+++ src/main/java/repositories/TrainerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import domain.Trainer;
 
 @Repository
-public interface TrainerRepository extends JpaRepository<Trainer, Long> {
+public interface TrainerRepository extends JpaRepository<Trainer, Integer> {
 
 }
